+++ requierments.txt
@@ -1,0 +1,3 @@
+kornia
+torch
+skimage
